requirements.txt
tgtg>=0.13.0
python-dotenv>=1.0.0
plyer>=2.1.0
